Use hierarchical RID for Collection synchronization
With HRIDs the CollectionSync can more easily and reliably detect RID
changes. This fixes problem during sync when Collection RID changes but
not its name. Previously the CollectionSync would not know it's still the
same Collection (just with different RID) and would first try to create
a new Collection, which would fail due to unique name-parent constraint.
With HRIDs CollectionSync is able to detect the RID change and issue a
CollectionModifyJob to update the RID, as expected.