Preserve icaldir resource/collection display name
Match the root collection's display name to AgentBase::name() and
update the agent name when user modifies the root collection's name.
This way it's possible to persistently customize icaldir resource
names instead of having multiple "Calendar Folder" entries in Korganizer.
It also makes the resource visible under the "nice" name in resource
listing, instead of using the agent identifier.

BUG: 376379
FIXED-IN: 5.4.3