Maildir: make use of the new Foreign Payload features from Akonadi
Now the data in the Item refer to the actual email file in the maildir
storage, so there's no data duplication between maildir and Akonadi
and the payload is always immediatelly available to clients without
them having to wait for the maildir resource to upload the payload
into Akonadi.

This is backwards compatible with existing Items that don't use
Foreign payloads, they will be converted once they expire and their
payload is requested again.

This also means that the cache policy can be changed to permanently
cache the entire Item, because there is no longer the overhead of
duplicated storage (which was the main reason why we were expiring
the bodies as soon as possible)