FB: Use a dedicated webengine profile and store cookies
We use a dedicated QtWebEngineProfile for each resource instance and
for each profile we store all cookies that Facebook creates during
login. This will allow us to query the webpage later to sync birthdays.
It also allows us to inject the cookies back into the profile before
refreshing token so that user does not have to re-enter their password
every time.