FB: work around FB API limitation to sync birthdays
With Graph API 2.0 apps can only see friends that also use the same app.
This is not very useful, since for most users the birthday calendar would
be mostly empty. Luckily Facebook provides an iCal with all friends'
birthdays, so we use a bit of cookie magic to download user's Facebook
Events page, parse the iCal URL from it and then fetch the iCal with all
birthdays. Yay for us.