Make KGAPI and SASL2 a hard dependency of kdepim-runtime
KPimGAPI is now released as part of KDE Applications and follows
the versioning of KDE Applications (and of KDE PIM, internally)
so there's no reason for it to be an optional dependency anymore.
I'm sure most ditributions ship it already anyway in order to
provide Google Calendar/Contacts support.

SASL2 is needed for the native GMail authentication support. Since
other KDE PIM libraries already have a hard dependency on SASL2
(like KIMAP) SASL2 is already present on all installations and so
we just make sure that all distros now ship with the Gmail XOAuth
support enabled.