EWS/test: Increase isolated test timeouts
The previous timeouts didn't take into account that the host system can
be severely loaded. This can cause the test to be aborted with a
misleading message that the resource hash crashed, which is caused by
the resource receiving a SIGTERM from the test suite, which is tearing
everything down after a timeout.