EWS: Remove inbox filtering workaround
This was needed until recently when D3181 was committed. Since its no
longer needed it can be removed.

The only thing left is the part that performs the reverse renaming from
INBOX to the actual Exchange ID for the Inbox folder. This is necessary
to bring existing EWS resource instances back to proper state.