Handle Error in GetFolderResponseMessage
Summary:
In case a Folder which no longer exists is updated, the server responds
with a GetFolderResponseMessage with ResponseClass="Error". This happens
if the internal cache still contains a folder, but the folder has been
deleted meanwhile (e.g. through the web client).

The message contains an empty Folders element which has to be skipped,
otherwise the outer loop for the GetFolderResponseMessages will try to
traverse the (nonexisting) Folders element on readNextStartElement.

BUG: 387580

Test Plan:
Create a new folder (either using web client or kontact)
Make sure it appears in kontact
Delete it in the web client (or using kontact on another computer)
Try to sync the resource

Reviewers: dvratil, nowicki

Reviewed By: dvratil

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12475