resources/ews: Fix saving passwords to KWallet
This is a regression after switching to asynchronous wallet
opening. When saving the password the wallet is opened and a callback is
registered to write the password once the wallet is ready. Unfortunately
all this is done in an instance of the EwsSettings class that is local
to the configuration dialog class. This means that this instance is
destroyed once the configuration dialog is dismissed, so when opening
the wallet takes a split second too long the object dies before the
callback has a chance to do its job.

The fix is not to use a local instance of the EwsSettings class in the
configuration dialog and instead work on a pointer to this object kept
by the root resource class, which is alive at all times.

BUG: 393002