Enable OAuth2 in EWS request
When a request is started in case of enabled OAuth2 the implementation
checks if a token is available and if not - launches the
authentication. The actual HTTP request is delayed until that process
is completed.

This process is also started in case a 401 Unauthorized HTTP response
is received with OAuth2 enabled.