EWS/test: Fix authentication failure in isolated test
Commit a20701f58 added the "has domain" flag to fix ambiguity between
empty domain and no domain in the username. Unfortunately the default
behaviour was set to "domain present", which changed the default
username from "test" to "/test" causing an unexpected value of the
Authorization HTTP header and in turn a failure.

Fix this by explicitly setting the "has domain" flag to false.