EWS/test: Work around occasional missed collection change notifications
The test works by faking responses to some EWS requests and waiting
for the resource to reach an expected state based on mocked responses
to these requests.

The state expectation logic is driven by Akonadi collection change
monitor. Unfortunately every now and then some notifications get lost
causing the state not to be reached while the server is actually in a
correct one.

Work around this by introducing an optional recheck method that
fetches all collections and forces a recheck of the current state.

This should finally fix the isolated tests for good.