pop3/autotests: wait for MaildirResource::setPath() to take effect
Setting the maildir path is an asynchronous operation of which effects
can be delayed. Fetching the maildir root collection before that will
cause the fetched identifier to become stale and cause a cascade of
failures.

This makes use of the internal information that the remote identifier
of the maildir resource is the path to the storage directory.