Qualify all POP3 mails for deletion if supplied id == -1
When determining ids of mails to delete a special identifier -1 is
used to reference all e-mails. When the "leave on server" flag is not
set asking for deletion of all mails should really return a list of
all mail ids.

This fixes Pop3Test::testSimpleLeaveOnServer test case.

Reviewers: #kde_pim

Reviewed By: dvratil

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D15962