Wait for Akonadi to process items after handling all POP3 messages
The POP3 resource test simulates a series of POP3 messages to the
resource and waits for them to be processed. Unfortunately this is not
enough as even though the POP3 resource is done talking to the server
processing of items in Akonadi and in the associated maildir resource
can still take some time. Due to this fact the POP3 tests were failing
intermittently.

In order to prevent that run an additional wait loop with the help of
the monitor interface. Each time an item operation is notified a timer
is reset. Once no item operation happens within 200ms it is safe to
assume that the server is done and all the operations resulting from
the simulated POP3 traffic have been committed.