Reenable OAuth2 on Qt < 5.10
The QtNetworkAuth module from Qt 5.9 does not give explicit access to
the refresh token. This means that there is no possibility to save or
restore the refresh token from the wallet, but it is possible to use
it while the resource is running.

The penalty is that on Qt 5.9 the user will have to reauthenticate
more often in case they restart their system. I still think it is
better to enable OAuth2 for such users with such drawback than to
completely take away the possibility to use modern authentication.