Google Contacts: fix Other Contacts collection
For some time Akonadi Server forces 'virtual' flag on sub-collections
that are children of a Collection that only permits virtual children
through contents mimetype. As a result of that the 'Other Contacts'
was created as a virtual collection, because the top-level collection
was only specifying the virtual collections mimetype.

The fix here is to allow non-virtual collections here as well. However,
due to this we need to disallow creation of new sub-collections, because
e cannot force the 'virtual' flag in KAddressBook at creation. In other
words, new contacts groups can only be created through web interface,
until a better solution is found.

Together with a recent change in CollectionSync that will now re-create
a collection if its 'virtual' flag has changed, the broken 'Other
Contacts' collection will correct itself on next sync.