Fix error handling in incremental folder sync job
The comment about error handling that was originally included was a
blind copy&paste from EwsFetchFoldersJob, where the initial sync can
fail due to overload, which is handled by falling back to an
identifier-only sync.

In case of EwsFetchFoldersIncrJob there is no such fallback, so the
error from the request must be propagated to the parent
request. Failure to do so can cause the parent request to end without
returning a result, which in turn can cause the resource state machine
to be stuck on a pending sync job.

In the past such problem almost never occurred, but nowadays with
OAuth2 authentication every request can randomly fail due to auth
error, so the problem became visible.