Skip update notifications of own modified items
Record items that have been updated locally and sent to the server so
that when the server notifies us back about their update the
notification can be skipped as the local copy is already in the
correct version.

This prevents redundant item sync requests.

Similar mechanism already exists for item deletion.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>