Fix unnecessary reauthentication prompts (try 2)
The original fix added a countermeasure to detect if reauthentication
is in progress and in such case avoid retrying it in case of failed
requests as it would confuse the state machine. Unfortunately the
initial state value was set to 0, which would still cause the above
countermeasure to fail in the initial state of the reauthemtication
procedure.

To fix this make the value of the initial state 1-based to fully
ensure that unwanted retriggering does not happen.

Additionally the state variable was implemented as an enum instead of
a plain integer.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>