Dynamically increment reconnect timeout
In case of an error leading to bringing the resource offline
temporarily it needs to periodically attempt to restart in case the
issue is intermittent.

Initially retry fairly quickly (15s) to catch some short intermittent
problems. If that doesn't help, extend the wait time longer and longer
in order not to hammer the server with requests.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>