Set deprecation versions to Qt 5.13 and KF 5.63 for the release branch
This avoids stable branches breaking while the world around them changes,
and unbreaks the release branch here in combination with KF 5.65. Laurent,
I think we should do that for all repos, do you happen to have a script
for that (as well as for the following merge that needs to ignore this)?

CCMAIL: montel@kde.org