Add initial server busy test
The EWS server uses a throttling mechanism to avoid abuse. This
mechanism can often engage when performing a full sync of a large
folder. This test aims at simulating this behaviour.

Initially a test case for the initial server connection is
implemented. Additional test cases will be added for occurence of this
error during different operations of the EWS resource.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>