Try estimate the progress of a full sync
This speculates that the sync is a full resync of an existing folder,
which already was synced before and the number of local items already
known will be approximately the same as the remote item count.

A true value could be obtained by issuing a GetFolder request before,
but this would mean an extra request just for UI purpose, which is
really not that important.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>