Implement chunked item retrieval
This splits item retrieval into 10-item chunks instead of retrieving
all of them at once. The goal is to maintain responsiveness for large
number of items and provide the user with progress information. This
also mitigates EWS server limitation as to the maximum number of items
to be retrieved using a single request.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>