[resources/ews] Adjust name of root collection to the name of the resource
Summary:
The name of root collection inside the Akonadi DB (and thus, i.e. inside KMail)
is always `akonadi_ews_resource#`, even if the resource itself was renamed.
Even though we call `mRootCollection->setName(name())`, the name get's
overriden after we fetch it from DB.
The settings UI changes the name of the resource, but not root collection.

This patch runs `CollectionModifyJob` to rename the root collection in both
scenarios: either when we fetch root collection or change the resource name.

BUG: 413078
BUG: 416663

Test Plan:
1) Open KMail -> Settings -> Accounts -> Receiving -> EWS account
2) Change "Account Name", save
3) The change gets applied to Mailbox View inside KMail UI immediately

Reviewers: dvratil, nowicki

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D27817