Register IMAP D-Bus interfaces before the service gets registered
Summary:
The account wizard accesses this right after the service is registered,
so we need to have this done in the ctor. I suspect this was done delayed,
as settings() is virtual and therefore calling that in the base ctor isn't
giving us the desired result. Calling it in every leaf ctor is ugly, but
given we want to phase out the Kolab resource anyway, this problem is going
away over time.

Reviewers: mlaurent

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D28583