POP3: Remove the singleton pattern from Settings
Summary:
It's causing issues with the "new" out-of-process settings
dialog which initializes a new Settings object each time and asserts
when the SettingsHelper Q_GLOBAL_STATIC is already initialized.

Instead just get rid of the singleton pattern and have Settings
object be owned by whomever created it and pass a reference to it
to the classes that need it.

BUG: 419726

Subscribers: mkoller, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D28732