Synchronize items directly in response to subscription notification
When encountering a server-side notification about changes in folders
issue a item sync request directly along with subsequent item
add/modify/delete requests instead of relying on Akonadi to trigger
this. This is possible as Akonadi will not notify the resource back
about changes it has done itself.

This saves some roundtrip time in the server-side driven
synchronization mechanism.