Be less strict when handling missing items during delete notifications
When handling a server-side notification about item removal it can
happen that the local item is already gone. In most cases this is the
result of the item actually having been deleted by us.

Even if this happened due to a sync failure between the server and
Akonadi, the item is now gone anyway from the server, so launching an
expensive full sync will not bring it back in any case.