Drop deprecated queued updates mechanism
This mechanism was introduced to cancel the feedback caused by
server-side notifications in response to item update or removal by the
EWS resource itself.

In the current implementation such bounced item updates are handled by
comparing the item remote revision with the change id. Bounced
deletions are ignored in case they don't exist locally any more as
when they no longer exist on the server there is really not much that
can be done to recover.