Also set Internal Date in ChangeItemTask::doStart AppendJob
When filtering messages in an IMAP resource with the Pipe through
action, the internal date of filtered messages is lost. This
appears to be due to a forgotten setInternalDate for the AppendJob
in ChangeItemTask::doStart. This patch adds the setInternalDate
call. It brings ChangeItemTask::doStart in line with
AddItemTask::doStart (additemtask.cpp:56) and
ReplaceMessageJob::start (replacemessagejob.cpp:36).

It fixes one issue reported in Bug 417206 and perhaps also the
other (original) one mentioned there.