Don't clear sync state if sync error is not related to the sync itself
Quite often the sync can fail due to trivial errors like expiration of
the authentication token or exceeded request quota on the EWS
server. In such case forcing clearing of the sync state is not needed
and results in unnecessary load.

To fix this, the EWS requests can now propagate the errors received
from the EWS server so that the controller is able to determine the
reason for the failure and act accordingly.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>