Fix linkage of EWS tests
The resource is pulled into a static library, which is then linked
into each of the test. The problem is that the linker will only pull
objects, for which symbol dependencies exist. This is not the case for
Qt resources, which by itself do not export anything and only rely on
static globals. This means that objects containing Qt resources need
to be explicitly named on the linker command line - otherwise they are
not linked in and therefore missing from the final executable.

Fixes: 3c966d602 (Use autorcc directly)
Signed-off-by: Krzysztof Nowicki <krissn@op.pl>