Consider the online state when attempting to reconnect
There's actually a comprehensive error condition handling in the method
above which properly distinguishing between transient and persistent
problems, but we just ignore that decision here and continuously
reconnect.

Together with https://invent.kde.org/pim/kimap/-/merge_requests/9 this
fixes the infinite SSL error dialog loop when rejecting to ignore an SSL
error to a large extend. You still get the dialog twice now, and then
after a few minutes again as this is considered to be a transient error
(e.g. caused by capture portals). This at least gives you the opportunity
now to actually fix the configuration or remove the resource.

CCBUG: 423424

(Bug 423424 remains open for SMTP)