Use DBus activation in NewMailNotifierShowMessageJob
KIO::ApplicationLauncherJob unconditionally executes the kmail binary, even if
KMail is already running. Not only is this slower, it also has the negative
effect of opening the main window if it's not open already, sometimes even in
front of the message that should be shown.

(cherry picked from commit 31b3194ae7041641e688da09fbd8fcf392af7831)