SingleFileResource: add option to periodically update file
When change monitoring is not possible (i.e. remote URL calendars),
allow users to configure automatic periodic updates. While technically
this is possible to configure manually in the generic "Folder Properties"
dialog for the calendar Akonadi Collection, users are unlikely to look
for the option there. Also most other resources have the periodic or
interval check option present in their configuration, which makes it
more likely for users to expect it even in the SingleFileResource-based
resource configuration dialogs.

This affects the iCal, Notes, MBox and VCard resources.

BUG: 428665
FIXED-IN: 23.12