Don't destroy save password job before closing the save dialog
Use an evil nested event loop for this as the agent save api is
synchronous and we need to stop the execution of it until we finish
saving the password.

Otherwise the complete account configuration won't be saved when
clicking on the Okay button but will be saved when clicking the Apply
button.