Remove Akonadi Relations from Kolab Resource
Relations, in Kolab terminology, include also Tags (or rather relation
between tag and an item) - this is preserved and still functional (probably),
only support for inter-item relations (which maps to Akonadi Relations)
has been removed. The support for propagating relations created in client
applications was there, but doesn't seem to have been called from anywhere,
so even in Kolab Resource, it only was a one-way sync. One more reason
to kill-off Akonadi Relations for good.