Do not try to subscribe to an empty list of folders
The server refuses to watch when the list of folders is empty, which
caused EwsSubscribedFoldersJob to try and subscribe in a loop. Just
subscribe to the default folder list when this happens so we avoid the
error. Subscribing still fails with a timeout, but at least we're not
trying to do something invalid.

CCBUG: 485799


(cherry picked from commit 46ad0976a178a0189fe9fea6ec99f4a8a15b5af3)

Co-authored-by: Louis Moureaux <m_louis30@yahoo.com>