Google: sync tags to Google Calendar
Google Calendar doesn't have any concept of tags or categories (you
can assign different colors to events, but that's not really tagging),
so we store the tags in event's ExtendedProperties. Those are then
read again on sync by other instance of Google Resource.

The key is not namespaced, so in theory any other client that will
look for 'categories' in the extended properties can support them
out-of-the-box.